Move storage package out from ingestor package
Used by collector and ingestor.
