+++ cmd/collector/main.go
@@ -21,9 +21,9 @@ import (
 	"github.com/Azure/adx-mon/collector/logs/sources/tail"
 	"github.com/Azure/adx-mon/collector/logs/transforms"
 	"github.com/Azure/adx-mon/collector/logs/types"
-	"github.com/Azure/adx-mon/ingestor/storage"
 	"github.com/Azure/adx-mon/pkg/k8s"
 	"github.com/Azure/adx-mon/pkg/logger"
+	"github.com/Azure/adx-mon/storage"
 	"github.com/pelletier/go-toml/v2"
 	"github.com/urfave/cli/v2"
 	"k8s.io/client-go/kubernetes"
