+++ ingestor/adx/syncer.go
@@ -10,8 +10,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/Azure/adx-mon/ingestor/storage"
 	"github.com/Azure/adx-mon/pkg/logger"
+	"github.com/Azure/adx-mon/storage"
 	"github.com/Azure/adx-mon/transform"
 	"github.com/Azure/azure-kusto-go/kusto"
 	"github.com/Azure/azure-kusto-go/kusto/unsafe"
