+++ tools/otlp/logs/transport_test.go
@@ -14,8 +14,8 @@ import (
 	ingestorsvc "github.com/Azure/adx-mon/ingestor"
 	"github.com/Azure/adx-mon/ingestor/adx"
 	ingestorotlp "github.com/Azure/adx-mon/ingestor/otlp"
-	"github.com/Azure/adx-mon/ingestor/storage"
 	"github.com/Azure/adx-mon/pkg/otlp"
+	"github.com/Azure/adx-mon/storage"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
